requirements.txt
elasticsearch>=7.8.0
aiohttp>=3.7.4.post0
fastapi>=0.63.0
python-dotenv>=0.15.0
uvicorn>=0.13.4

